requirements.txt
Package            Version
------------------ ------------
amqp               5.2.0
APScheduler        3.10.4
async-timeout      4.0.3
billiard           4.2.0
blinker            1.7.0
certifi            2023.11.17
charset-normalizer 3.3.2
click              8.1.7
click-didyoumean   0.3.0
click-plugins      1.1.1
click-repl         0.3.0
Flask              3.0.1
Flask-APScheduler  1.13.1
huey               2.5.0
idna               3.6
itsdangerous       2.1.2
Jinja2             3.1.3
kombu              5.3.5
logger             1.4
MarkupSafe         2.1.4
numpy              1.26.3
pandas             2.2.0
pip                23.3.1
prompt-toolkit     3.0.43
python-dateutil    2.8.2
python-dotenv      1.0.1
pytz               2023.3.post1
redis              5.0.1
requests           2.31.0
setuptools         68.2.2
six                1.16.0
spotipy            2.23.0
SQLAlchemy         2.0.25
typing_extensions  4.9.0
tzdata             2023.4
tzlocal            5.2
urllib3            2.1.0
vine               5.1.0
wcwidth            0.2.13
Werkzeug           3.0.1
wheel              0.41.3
